Strangler extraction #13: config-file I/O -> zxnu_config_io.py
Move load_configuration_file (the whole hdfg.cfg restore pipeline that
pushes every saved setting back into the constructed widget tree) and
save_configuration_file (the _initialising-guarded UTF-8 writer) out of
MainWindow.__init__ into build_config_io(host, ...).

Same builder seam: self -> host via a tokenize NAME pass; helpers the
restore calls that are defined later in __init__ injected as forwarding
lambdas; ``wid_inner`` and the module-global _ZXNU_START_RE_LISTENER
flag read via getter hooks at call time (the block is definition-only,
both closures run only after startup construction). Both functions are
written to host and re-bound to bare __init__ locals at the call site,
re-establishing the closure cell the rest of __init__ captures.
zxnu_main.py 9,467 -> 8,580 lines. Tripwired in test_pane_imports.py.
